backend/requirements.txt
fastapi
gpt4all
uvicorn
pydantic
httpx
requests
python-dotenv
langchain-community
faiss-cpu
sentence-transformers
